filter-rule-table: allow for non-integral rule counts.
git-svn-id: https://mosesdecoder.svn.sourceforge.net/svnroot/mosesdecoder/trunk@4036 1f5c12ca-751b-0410-a591-d2e778427230

+++ scripts/training/filter-rule-table.py
@@ -119,7 +119,7 @@ def parseRule(line):
     if len(cols) > 4:
         counts = cols[4].split()
         if len(counts) == 3:
-            ruleCount = int(counts[2])
+            ruleCount = float(counts[2])
     return (rhsSourceSymbols, ruleCount)
 
 def isNT(symbol):
